ai-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
numpy==2.1.3
faiss-cpu==1.10.0
pillow==11.0.0
transformers==4.52.4
torch==2.7.1
diffusers==0.33.1
accelerate==1.8.1
